include nuid in stat collection dedupe
Some networks (gira) generate a ghost station while it's broken. Other
systems have multiple stations on the same exact latitude and
longitudes. For stats deduping we need to include an internal uid of the
station. Not all systems have internal uid (nuid) though, so this works
on a best effort basis.